GradAttack/requirements.txt
colorama==0.4.4
matplotlib==3.4.3
#numpy==1.21.4
opacus==0.14.0
pytorch_lightning==1.5.1
scikit_learn==1.0.1
#torch==1.8.1
#torchcsprng==0.2.1
torchmetrics==0.6.0
torchvision==0.9.1
dill
